pyproject.toml
[project]
name = "giros-bot"
version = "0.1.0"
description = "Sistema de generación de contenido automatizado (Blog & RRSS Omnicanal) para Giros Media SpA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # IA / LangGraph
    "langgraph>=0.3.0",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.0.0",
    # Validación
    "pydantic>=2.10",
    "pydantic-settings>=2.6.0",
    # GitHub
    "pygithub>=2.5.0",
    # HTTP
    "httpx>=0.27.0",
    # YAML (frontmatter MDX)
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.0",
    "tavily-python>=0.7.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/giros_bot"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
